Update Prototype_Code.ino
Proof of concept code. This code checks the DC voltage that gets through each bandpass filter. We used 3 bandpass filters with center frequencies of 60 Hz, 400 Hz, and 3.3 kHz. The output of each filter is connected to their respective input pin on the Arduino Micro. So far, when the voltage is greater than 0 for one of the bands, the program will send the matrix code out. There is still a lot to figure out in regard to the code, filters, and microphone amplification.